refactor(direct-control): extract tactical read types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only settlement recommendation input/factor/origin/result types into packages/civ7-direct-control/src/play/tactical/settlement.ts, target-candidates input/candidate/result types into packages/civ7-direct-control/src/play/tactical/target-candidates.ts, battlefield scan input/result types into packages/civ7-direct-control/src/play/tactical/battlefield.ts, and destination analysis input/result types into packages/civ7-direct-control/src/play/tactical/destination.ts. Keeps public facade type re-exports in index.ts.

Preserves public type contracts and does not change tactical source strings, command serialization, parser labels, read-only/no-send behavior, conservative relationship-label policy, runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Diplomacy/narrative closeout, operation, ready, public procedure schema, and telemetry type ownership remain pending separate owner slices.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- tactical-reads.test.ts settlement-recommendations.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.play.tactical-read.test.ts game.play.settlement-recommendations.test.ts game/play/priorities.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, tactical read behavior, relationship classification, CLI semantic projection, telemetry, AI-ingestion, hotseat product-path, or procedure-core proof claimed.
